Drop support for go 1.7
The codegen routines for ksonnet-lib depend on some Go 1.8 features
(e.g., `sort.Slice`). Once we vendor the dependency, this will cause the
build to fail perpetually for Go 1.7.

This commit therefore drops support for 1.7, making 1.8 the line of
supportability. This decision was made together with @anguslees, in
slack PMs.