Construct base components object
In order to support environment heirarchy, we need to be able to
reference all components. This commit will implement component imports
as k-v pairs ex: foo: "import/foo.jsonnet" as a Jsonnet object. This
jsonnet object will then be assigned as an ExtCode so that it can be
referenced by a base.libsonnet file which environments can build /
override upon.