Remove unnecessary param from defaultNamespace()
The `clientConfig` param currently being passed is not needed, because
it exists as a package level var in root. It also makes little sense to
pass a custom `clientConfig` because if `overrides.Context.Namespace` is
populated, the namespace that is returned is configured as an override
in the package level `clientConfig` and not the `clientConfig` in the
param.