🆘 Add `--help` flag to `generate` and `prototype use`
When `generate` was first implemented, we made a conscious decision to
not include the `--help` flag, so that we would not accidentally break
templates with a parameter called `help`.

But, this is unintuitive for users who want to use the `--help` flag, so
in this commit we will print the help message if (1) there is a single
command argument, and (2) it is either `--help` or `-h`.

Fixes #21.