📦 Rename command `dep` -> `pkg`
The command `dep` currently broadly allows users to inspect, use, and
search for packages. Currently, however, the name `dep` suggests that
this functionality is scoped to dependencies. Consider, for example, the
command `dep list`. A user might reasonable expect this to list all
dependencies, rather than all packages, which is what it does. Another
example is `dep search`, which actually searches all known packages, not
all known dependencies.

Hence, this commit will rename this command to simply be `pkg`.