Rewrite integration tests to support ksonnet.next application structure
ksonnet.next is centered around the application directory structure
generated during init, however the integration tests are still tightly
tied to the model from kubecfg, where the '-f' flag was the standard and
environments were not introduced.

This commit will add test fixtures to mock a ksonnet application
structure, and rewrite the existing apply and delete tests to support
this new model.

Signed-off-by: Jessica Yuen <im.jessicayuen@gmail.com>