Validate app, parts, and registry schemas on deserialization
Dependencies in a ksonnet application are installed from collections of
packages called registries. In order to be precise about which packages
an app depends on, and where they are from, we have an API that
specifies each of these nouns.

Unfortunately, we never hooked up code to verify the versions of these
specifications as we deserialize them.

This commit will correct that.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>