App spec to take a single destination
Currently the app.yaml spec takes a list of destinations. The change
exists to anticipate support for an environment supporting multiple
clusters or "destinations". The problem is, while the use case makes
sense for `apply`, and `delete`, it becomes ambiguous which cluster is
being referred to during commands such as `diff`. We've considered
specifying a cluster during a `diff`, however, ks currently doesn't have
a notion of a cluster identity. This change is to update the app.yaml to
take a single destination to more accurately represent the state of
things.

Signed-off-by: Jessica Yuen <im.jessicayuen@gmail.com>