Fix bug with invalid base.libsonnet import path
There exists a bug in the reference path of base.libsonnet for nested
environment directories. We are referencing a hardcoded import
"../base.libsonnet". The works for top level environment directories but
obviously does not work for nested directories.

This commit will add the environments directory as a global lib path.

Signed-off-by: Jessica Yuen <im.jessicayuen@gmail.com>