edit ksonnet license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit renames license_test.go to apache_header_test.go.
Prior to this change, Licensee detected the word "license" in the test
filename and mistook it for a license file of unknown type. Even though
Licensee correctly identified LICENSE as an Apache license, its
detection of license_test.go added ambiguity to its reporting of the
project's license type.

Removing the source of ambiguity by renaming the file allows Licensee
to successfully identify the license type of ksonnet as Apache.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>