adding component selection for diff
This change adds components for diffs. You can now run:

```
ks diff local:default remote:default -c <component name>
```

To diff a component in a module, specify the component with a module name. e.g component (for a module in the root component) or nested.component (for a module in nested/ under components)

In addition, module support throughout the ks was updated

* Modules allow ksonnet users to separate their configurations under components.
* Modules are dot separated paths that correlate to their location on the file system. e.g module a is in /app/components/a
* Modules can be deeply nested. e.g components in /app/components/nested/deeply would be in module nested.deeply
* Modules are now added as a label to kubernetes resources in `ksonnet.io/component`. This will allow the user to determine which component created a resource

Signed-off-by: bryanl bryanliles@gmail.com